+++ R/SVGElement.R
@@ -102,6 +102,14 @@ SVGElement <- R6::R6Class(
       do.call(self$append, children)
 
       #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+      # If duplicate named attributes are given, keep only the last one
+      #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+      dups <- rev(duplicated(rev(names(attribs))))
+      if (any(dups)) {
+        attribs <- attribs[!dups]
+      }
+
+      #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
       # If one of the values for an attribute is an SVGElement, convert it
       # into an ID
       #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~