+++ NEWS.md
@@ -1,12 +1,17 @@
 
-# minisvg 0.1.2  20200112
+# minisvg 0.1.3  2020-01-18
+
+* When multiple attributes with the same name are given for an element, then 
+  only the last one is kept/used.
+
+# minisvg 0.1.2  2020-01-12
 
 * For `animate` and related commands, if vectors are passed in as the value
   for a `values` or `keyTimes` argument, then `paste(x, collapse=";")` is applied. 
   This will make some animation setup easier from R since we no longer have 
   to collpse vectors before calling.
 
-# minisvg 0.1.1  20200110
+# minisvg 0.1.1  2020-01-10
 
 * Added `feBlend` and other filter effects to `stag` and `svg_elem`
 * Added vignetted on `filter-elements`